requirements.txt
ultralytics
deep_sort_realtime
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-huggingface==0.1.2
langchain-nvidia-ai-endpoints==0.3.9
langchain-openai==0.3.9
langchain-text-splitters==0.3.6
langgraph==0.3.21
langgraph-api==0.0.38
langgraph-checkpoint==2.0.23
langgraph-checkpoint-sqlite==2.0.6
langgraph-cli==0.1.81
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langsmith==0.3.13
numpy==1.26.4
openai==1.66.3
pandas==2.1.4
pillow==11.1.0
PyAudio==0.2.14
pydantic==2.10.6
pyttsx3==2.98
pytz==2025.1
pywin32==310
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==3.4.1
setuptools==76.1.0
SpeechRecognition==3.14.1
SQLAlchemy==2.0.39
sympy==1.13.1
tavily-python==0.5.3
torch==2.6.0
tornado==6.4.2
transformers==4.49.0
typing_extensions==4.12.2
uvicorn==0.34.0
wikipedia==1.4.0
stereo_camera_service

